refactor: extract item frame place related events to independent listener
