+++ .editorconfig
@@ -13,7 +13,7 @@ ij_formatter_off_tag = @formatter:off
 ij_formatter_on_tag = @formatter:on
 ij_formatter_tags_enabled = true
 ij_smart_tabs = false
-ij_visual_guides = 
+ij_visual_guides =
 ij_wrap_on_typing = true
 
 [*.java]
@@ -67,7 +67,7 @@ ij_java_blank_lines_before_package = 0
 ij_java_block_brace_style = end_of_line
 ij_java_block_comment_add_space = false
 ij_java_block_comment_at_first_column = true
-ij_java_builder_methods = 
+ij_java_builder_methods =
 ij_java_call_parameters_new_line_after_left_paren = false
 ij_java_call_parameters_right_paren_on_new_line = false
 ij_java_call_parameters_wrap = off
@@ -88,7 +88,7 @@ ij_java_doc_add_blank_line_after_return = false
 ij_java_doc_add_p_tag_on_empty_lines = true
 ij_java_doc_align_exception_comments = true
 ij_java_doc_align_param_comments = true
-ij_java_doc_do_not_wrap_if_one_line = false
+ij_java_doc_do_not_wrap_if_one_line = true
 ij_java_doc_enable_formatting = true
 ij_java_doc_enable_leading_asterisks = true
 ij_java_doc_indent_on_continuation = false
@@ -101,32 +101,32 @@ ij_java_doc_param_description_on_new_line = false
 ij_java_doc_preserve_line_breaks = false
 ij_java_doc_use_throws_not_exception_tag = true
 ij_java_else_on_new_line = false
-ij_java_entity_dd_prefix = 
+ij_java_entity_dd_prefix =
 ij_java_entity_dd_suffix = EJB
-ij_java_entity_eb_prefix = 
+ij_java_entity_eb_prefix =
 ij_java_entity_eb_suffix = Bean
-ij_java_entity_hi_prefix = 
+ij_java_entity_hi_prefix =
 ij_java_entity_hi_suffix = Home
 ij_java_entity_lhi_prefix = Local
 ij_java_entity_lhi_suffix = Home
 ij_java_entity_li_prefix = Local
-ij_java_entity_li_suffix = 
+ij_java_entity_li_suffix =
 ij_java_entity_pk_class = java.lang.String
-ij_java_entity_ri_prefix = 
-ij_java_entity_ri_suffix = 
-ij_java_entity_vo_prefix = 
+ij_java_entity_ri_prefix =
+ij_java_entity_ri_suffix =
+ij_java_entity_vo_prefix =
 ij_java_entity_vo_suffix = VO
 ij_java_enum_constants_wrap = off
 ij_java_enum_field_annotation_wrap = off
 ij_java_extends_keyword_wrap = off
 ij_java_extends_list_wrap = off
 ij_java_field_annotation_wrap = split_into_lines
-ij_java_field_name_prefix = 
-ij_java_field_name_suffix = 
-ij_java_filter_class_prefix = 
-ij_java_filter_class_suffix = 
-ij_java_filter_dd_prefix = 
-ij_java_filter_dd_suffix = 
+ij_java_field_name_prefix =
+ij_java_field_name_suffix =
+ij_java_filter_class_prefix =
+ij_java_filter_class_suffix =
+ij_java_filter_dd_prefix =
+ij_java_filter_dd_suffix =
 ij_java_finally_on_new_line = false
 ij_java_for_brace_force = never
 ij_java_for_statement_new_line_after_left_paren = false
@@ -136,7 +136,7 @@ ij_java_generate_final_locals = false
 ij_java_generate_final_parameters = false
 ij_java_generate_use_type_annotation_before_type = true
 ij_java_if_brace_force = never
-ij_java_imports_layout = *,|,javax.**,java.**,|,$*
+ij_java_imports_layout = *, |, javax.**, java.**, |, $*
 ij_java_indent_case_from_switch = true
 ij_java_insert_inner_class_imports = false
 ij_java_insert_override_annotation = true
@@ -161,13 +161,13 @@ ij_java_layout_static_imports_separately = true
 ij_java_line_comment_add_space = false
 ij_java_line_comment_add_space_on_reformat = false
 ij_java_line_comment_at_first_column = true
-ij_java_listener_class_prefix = 
-ij_java_listener_class_suffix = 
-ij_java_local_variable_name_prefix = 
-ij_java_local_variable_name_suffix = 
-ij_java_message_dd_prefix = 
+ij_java_listener_class_prefix =
+ij_java_listener_class_suffix =
+ij_java_local_variable_name_prefix =
+ij_java_local_variable_name_suffix =
+ij_java_message_dd_prefix =
 ij_java_message_dd_suffix = EJB
-ij_java_message_eb_prefix = 
+ij_java_message_eb_prefix =
 ij_java_message_eb_suffix = Bean
 ij_java_method_annotation_wrap = split_into_lines
 ij_java_method_brace_style = end_of_line
@@ -182,17 +182,17 @@ ij_java_new_line_after_lparen_in_annotation = false
 ij_java_new_line_after_lparen_in_deconstruction_pattern = true
 ij_java_new_line_after_lparen_in_record_header = false
 ij_java_new_line_when_body_is_presented = false
-ij_java_packages_to_use_import_on_demand = java.awt.*,javax.swing.*
+ij_java_packages_to_use_import_on_demand = java.awt.*, javax.swing.*
 ij_java_parameter_annotation_wrap = off
-ij_java_parameter_name_prefix = 
-ij_java_parameter_name_suffix = 
+ij_java_parameter_name_prefix =
+ij_java_parameter_name_suffix =
 ij_java_parentheses_expression_new_line_after_left_paren = false
 ij_java_parentheses_expression_right_paren_on_new_line = false
 ij_java_place_assignment_sign_on_next_line = false
 ij_java_prefer_longer_names = true
 ij_java_prefer_parameters_wrap = false
 ij_java_record_components_wrap = normal
-ij_java_repeat_annotations = 
+ij_java_repeat_annotations =
 ij_java_repeat_synchronized = true
 ij_java_replace_instanceof_and_cast = false
 ij_java_replace_null_check = true
@@ -203,23 +203,23 @@ ij_java_resource_list_wrap = off
 ij_java_rparen_on_new_line_in_annotation = false
 ij_java_rparen_on_new_line_in_deconstruction_pattern = true
 ij_java_rparen_on_new_line_in_record_header = false
-ij_java_servlet_class_prefix = 
-ij_java_servlet_class_suffix = 
-ij_java_servlet_dd_prefix = 
-ij_java_servlet_dd_suffix = 
-ij_java_session_dd_prefix = 
+ij_java_servlet_class_prefix =
+ij_java_servlet_class_suffix =
+ij_java_servlet_dd_prefix =
+ij_java_servlet_dd_suffix =
+ij_java_session_dd_prefix =
 ij_java_session_dd_suffix = EJB
-ij_java_session_eb_prefix = 
+ij_java_session_eb_prefix =
 ij_java_session_eb_suffix = Bean
-ij_java_session_hi_prefix = 
+ij_java_session_hi_prefix =
 ij_java_session_hi_suffix = Home
 ij_java_session_lhi_prefix = Local
 ij_java_session_lhi_suffix = Home
 ij_java_session_li_prefix = Local
-ij_java_session_li_suffix = 
-ij_java_session_ri_prefix = 
-ij_java_session_ri_suffix = 
-ij_java_session_si_prefix = 
+ij_java_session_li_suffix =
+ij_java_session_ri_prefix =
+ij_java_session_ri_suffix =
+ij_java_session_si_prefix =
 ij_java_session_si_suffix = Service
 ij_java_space_after_closing_angle_bracket_in_type_argument = false
 ij_java_space_after_colon = true
@@ -301,14 +301,14 @@ ij_java_spaces_within_synchronized_parentheses = false
 ij_java_spaces_within_try_parentheses = false
 ij_java_spaces_within_while_parentheses = false
 ij_java_special_else_if_treatment = true
-ij_java_static_field_name_prefix = 
-ij_java_static_field_name_suffix = 
-ij_java_subclass_name_prefix = 
+ij_java_static_field_name_prefix =
+ij_java_static_field_name_suffix =
+ij_java_subclass_name_prefix =
 ij_java_subclass_name_suffix = Impl
 ij_java_switch_expressions_wrap = normal
 ij_java_ternary_operation_signs_on_next_line = false
 ij_java_ternary_operation_wrap = off
-ij_java_test_name_prefix = 
+ij_java_test_name_prefix =
 ij_java_test_name_suffix = Test
 ij_java_throws_keyword_wrap = off
 ij_java_throws_list_wrap = off
